Update update-distro.ps1
Added user confirmation to approve the changes and added fail-safe for empty arrays

+++ update-distro.ps1
@@ -1,18 +1,38 @@
 param($distroList, $csv)
 
-$newUsers = @(Get-Content -Path $csv)
+$updatedUsers = @(Get-Content -Path $csv)
 
 $currentUsers = @((Get-DistributionGroupMember -Identity $distroList).PrimarySmtpAddress)
 
-$addUsers = @($newUsers | Where-Object {$currentUsers -NotContains $_})
-$removeUsers = @($currentUsers | Where-Object {$newUsers -NotContains $_})
+$addUsers = @($updatedUsers | Where-Object {$currentUsers -NotContains $_})
+$removeUsers = @($currentUsers | Where-Object {$updatedUsers -NotContains $_})
 
-foreach ($member in $addUsers) {
-    Add-DistributionGroupMember -Identity $distro -Member $member
-    Write-Host "Added" $member "to" $distro "Distribution List"
-}
+Write-Host "`nUsers to be Added:"
+$addUsers
+Write-Host "`nUsers to be Removed:"
+$removeUsers
 
-foreach ($member in $removeUsers) {
-    Remove-DistributionGroupMember -Identity $distro -Member $member
-    Write-Host "Removed" $member "from" $distro "Distribution List"
+$confirmation = Read-Host "`nDo you wish to continue with this process?? (Y/N)"
+
+if($confirmation -eq "Y" -or $confirmation -eq "y" -or $confirmation -eq "Yes" -or $confirmation -eq "yes" -or $confirmation -eq "YES") {
+    if([string]::IsNullOrEmpty($addUsers)) {
+        Write-Host "No users to add"
+    } else {
+        foreach ($member in $addUsers) {
+            Add-DistributionGroupMember -Identity $distro -Member $member
+            Write-Host "Added" $member "to" $distroList "Distribution List"
+        }
+    }
+
+    if([string]::IsNullOrEmpty($removeUsers)) {
+        Write-Host "No users to remove"
+    } else {
+        foreach ($member in $removeUsers) {
+            Remove-DistributionGroupMember -Identity $distro -Member $member
+            Write-Host "Removed" $member "from" $distroList "Distribution List"
+        }
+    }
+} else {
+    Write-Host "No Actions Taken"
+    break
 }
